Prune large diffs in DiffCollection
Diffs are pruned based on the total number of bytes, with the current
threshold being 100 KB. If a diff is larger than this threshold its data
is replaced with an empty String, but the Diff instance itself is
_still_ yielded in DiffCollection. This allows GitLab to still display
some data (e.g. the file path) _without_ trying to render a giant diff
(e.g.  minified CSS files).

The result of too_large? / prune_large_diff! is memoized so that GitLab
CE can store a boolean in the database and re-use this in combination
with the too_large? method to display a message about the diff size.