Handle collapsable DiffCollection
We want to give the user the opportunity to see more
files while preserving safe limits on rendering diffs
so files bigger than 10KB (we consider too large 100KB)
their lines are not taking into account to decide
if the collection overflow and you can decide if
you want to expand them on click or no.

To keep good performance we start collapsing all
the files since we are over the safe limits.

To do that you have to methods `collapsed?` and
`collapse` which means if it's quite large and the
other if the collection collapsed it when iterated
as a whole.

It's quite similar to the `too_large`concept but
give us a little bit more flexibility.