Add Repository#find_branch to speed up branch lookups
A common call in GitLab is to lookup a single branch, but previously
this was done by calling Repository#branches, which loads all the
branches into memory unnecessarily and causes many filesystem
accesses for each branch. With Repository#find_branch, we can
do a direct lookup for the branch we care about.